pyproject.toml
[project]
name = "rappimakers-dashboard"
version = "0.1.0"
description = "AI-Powered Dashboard for Rappi Store Availability"
authors = [
    {name = "Danilalo"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "pandas (>=2.0,<3)",
    "pyarrow (>=23.0.0,<24.0.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "plotly (>=6.5.2,<7.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-openai (>=1.1.9,<2.0.0)",
    "langchain-experimental (>=0.4.1,<0.5.0)",
    "streamlit (>=1.54.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "kaleido (>=1.2.0,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
